+++ source/include/libc/string.h
@@ -19,6 +19,6 @@ const char* finds(const char* src, const char* s);
 
 void reverse(char* s, int len);
 
-void itoa(int x, char* buffer);
+void itoa(uint64_t x, char* buffer);
 
 #endif
